Read e-Stat aggregation unit boundary data

The GIS data downloaded from e-Stat is read and converted into an easy to process sf::sf format. You can use download_stat_map() to download the data.

read_estat_map(file, type = "aggregate_unit", remove_cols = TRUE)

Arguments

file

Path to downloaded e-Stat shape file

type

Currently, only "aggregate_unit" is used.

remove_cols

Whether or not to remove redundant columns. When TRUE (the default), the following columns are removed (See details). These columns can be substituted or sought with values from other columns.

S_AREA

KAxx_, KAxx_id

KEN, KEN_NAME

DUMMY1

X_CODE, Y_CODE